(SDK-179) Start layering template directories
This commit has a provisional multi-layer layout for addressing the
different needs of `pdk new module` vs. `msync update`. It is intended to
use the moduleroot+puppetlabs directories for `msync update`, while the
moduleroot+new_module directories are for `pdk new module`. This structure
will allow fine-grained sharing of content without diverging too far from
the shareable commonalities.

This is for review only, as both pdk and msync do not work with this format
yet.